refactor: Corrige validation recherche avec localité
